refactor(api): Move old module package into legacy_api
