+++ api/opentrons/legacy_api/api.py
@@ -1,13 +1,15 @@
-from . import robot, instruments as inst, containers as cnt
+from . import robot, instruments as inst, containers as cnt, modules
 from .instruments import pipette_config
 
 # Ignore the type here because well, this is exactly why this is the legacy_api
 robot = robot.Robot()  # type: ignore
+modules.provide_singleton(robot)
 
 
 def reset():
     global robot
     robot = robot.Robot()
+    modules.provide_singleton(robot)
     return robot
 
 
@@ -259,5 +261,7 @@ class InstrumentsWrapper(object):
 instruments = InstrumentsWrapper(robot)
 containers = ContainersWrapper(robot)
 labware = ContainersWrapper(robot)
+modules.provide_labware(labware)
 
-__all__ = ['containers', 'instruments', 'labware', 'robot', 'reset']
+
+__all__ = ['containers', 'instruments', 'labware', 'robot', 'reset', 'modules']
